refactor(compiler): 重构编译器工厂并移除编译器注册表(阶段1)
遵循单一职责原则：现在 CompilerFactory 负责编译器的创建和管理，不再有多余的注册表层
简化架构：从三层结构（Factory + Registry + Service）简化为单层结构（Factory）
保持功能完整性：所有原有功能都被保留，只是重新组织了代码结构件
